Create nl.i18n.json
Dutch (Netherlands) translation file. In some cases I'm not sure what the right translation should be, because I don't know how a phrase is used in DE. For example the word "funds". Does it mean "finances" or is it more symbolic like tokens or goodwill or so? It would be good if an other Dutch person runs through the translation to look for wrong translations (comparing with the English file is the best way to do this I think). I would love to test the translation file in DE itself.